update 0.16.3v
- Update doc
- Update test
- Update route
- Fix bug
- Slight drop in smell code rate

+++ src/models/order_info.ts
@@ -42,11 +42,11 @@ export class OrderInfo extends Model {
         defaultValue: Sequelize.fn('now'),
       },
       sold_before_order: {
-        type: new DataTypes.INTEGER,
+        type: new DataTypes.FLOAT,
         allowNull: false,
       },
       total: {
-        type: new DataTypes.INTEGER,
+        type: new DataTypes.FLOAT,
         allowNull: false,
       },
       done: {
